+++ integration/flows/src/main/java/fr/unice/polytech/hcs/flows/car/g2/G2Car.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class G2Car {
-    @JsonProperty public int price;
+    @JsonProperty public Double price;
     @JsonProperty public String name;
 
     @Override
